mcpserver/requirements.txt
mcp==1.6.0
firecrawl-py==2.6.0
python-dotenv==1.0.1